+++ elog_backend/src/handlers/user_role_handler.rs
@@ -19,7 +19,7 @@ pub async fn get_user_roles(authenticated_request: AuthenticatedRequest) -> Resu
 #[get("/user_roles/{id}")]
 pub async fn get_user_role_with_id(
     authenticated_request: AuthenticatedRequest,
-    path: web::Path<(i8,)>
+    path: web::Path<(i16,)>
 ) -> Result<HttpResponse, ElogError> {
     Ok(HttpResponse::Ok().json(UserRole::get_with_id(&authenticated_request.connection, path.into_inner().0)))
 }
